refactor-settings * Delete ThemeProvider and move functionality into useThemeHandler-hook. * Use useThemeHandler on all tavle-types except Poster
