+++ src/containers/MyBoards/MyBoards.tsx
@@ -11,7 +11,7 @@ import {
 import { useUser } from '../../UserProvider'
 import { Board, SharedBoard, Theme } from '../../types'
 import { NoTavlerAvailable, NoAccessToTavler } from '../Error/ErrorPages'
-import { ThemeContrastWrapper } from '../ThemeWrapper/ThemeContrastWrapper'
+import { ThemeContrastWrapper } from '../ThemeContrastWrapper/ThemeContrastWrapper'
 import { Navbar } from '../Navbar/Navbar'
 import { SharedBoards } from './SharedBoards/SharedBoards'
 import { OwnedBoards } from './OwnedBoards/OwnedBoards'
